+++ src/cuda/sigmoid_ce.cu
@@ -187,7 +187,7 @@ torch::Tensor sigmoid_cross_entropy_forward(
     const int H_t = targets.size(1);
     const int W_t = targets.size(2);
 
-    const int total_elements = L * B * C * H_t * W_t;
+    const int64_t total_elements = (int64_t) L * B * C * H_t * W_t;
     if (total_elements == 0) {
         return torch::zeros({L}, logits.options());
     }
@@ -241,7 +241,7 @@ torch::Tensor sigmoid_cross_entropy_backward(
     const int W_t = targets.size(2);
 
     auto grad_logits = torch::empty_like(logits);
-    const int total_elements = L * B * C * H * W;
+    const int64_t total_elements = (int64_t) L * B * C * H * W;
     if (total_elements == 0) return grad_logits;
 
     float norm = num_masks / static_cast<float>(L);
